+++ src/terminal/Color.h
@@ -19,6 +19,7 @@
 #include <algorithm>
 #include <cstdint>
 #include <initializer_list>
+#include <optional>
 #include <ostream>
 #include <string>
 #include <variant>
@@ -164,8 +165,8 @@ struct ColorProfile {
 
     RGBColor defaultForeground = 0xD0D0D0;
     RGBColor defaultBackground = 0x000000;
-    RGBColor selection = 0x707070;
-    float selectionOpacity = 1.0f;
+    std::optional<RGBColor> selectionForeground = {0xb0b0b0};
+    std::optional<RGBColor> selectionBackground = {0x303030};
 	RGBColor cursor = 0x707020;
 
     RGBColor mouseForeground = 0x800000;